Use FP16 for all resamplers

+++ server/voice_changer/RVC/inferencer/RVCInferencer.py
@@ -49,5 +49,5 @@ class RVCInferencer(Inferencer):
             return_length=return_length,
             formant_length=formant_length
         )
-        res = res[0][0, 0].float()
+        res = res[0][0, 0]
         return torch.clip(res, -1.0, 1.0)
